[PresentationLayer] Activities renaming
