+++ presentation_layer/src/main/java/com/nexaas/challenge/presentation/core/ModulesInjection.kt
@@ -3,8 +3,8 @@ package com.nexaas.challenge.presentation.core
 import com.nexaas.challenge.data.remote.ApiRepositoryImpl
 import com.nexaas.challenge.domain.interactor.GetProductsList
 import com.nexaas.challenge.domain.repository.ApiRepository
-import com.nexaas.challenge.presentation.view.list.ProductsListPresenter
-import com.nexaas.challenge.presentation.view.product.ProductPresenter
+import com.nexaas.challenge.presentation.view.cart.CartPresenter
+import com.nexaas.challenge.presentation.view.productdetails.ProductDetailsPresenter
 import org.koin.dsl.module
 
 /* Objects common to the entire application */
@@ -29,7 +29,7 @@ val domainModule = module {
  * Presentation Module
  */
 val presentationModule = module {
-    factory<ProductsListPresenter> { ProductsListPresenter(get()) }
-    factory<ProductPresenter> { ProductPresenter() }
+    factory<CartPresenter> { CartPresenter(get()) }
+    factory<ProductDetailsPresenter> { ProductDetailsPresenter() }
 }
