+++ presentation_layer/src/main/java/com/nexaas/challenge/presentation/view/cart/CartView.kt
@@ -1,0 +1,7 @@
+package com.nexaas.challenge.presentation.view.cart
+
+import com.nexaas.challenge.presentation.core.mvp.BaseView
+
+internal interface CartView: BaseView {
+    fun updateList()
+}
