--- presentation_layer/src/main/java/com/nexaas/challenge/presentation/view/list/ProductsListView.kt
@@ -1,7 +1,0 @@
-package com.nexaas.challenge.presentation.view.list
-
-import com.nexaas.challenge.presentation.core.mvp.BaseView
-
-internal interface ProductsListView: BaseView {
-    fun updateList()
-}
