--- presentation_layer/src/main/java/com/nexaas/challenge/presentation/view/product/ProductPresenter.kt
@@ -1,6 +1,0 @@
-package com.nexaas.challenge.presentation.view.product
-
-import com.nexaas.challenge.presentation.core.mvp.BasePresenter
-
-internal class ProductPresenter: BasePresenter<ProductView>() {
-}
