--- presentation_layer/src/main/java/com/nexaas/challenge/presentation/view/product/ProductView.kt
@@ -1,6 +1,0 @@
-package com.nexaas.challenge.presentation.view.product
-
-import com.nexaas.challenge.presentation.core.mvp.BaseView
-
-internal interface ProductView: BaseView {
-}
