+++ presentation_layer/src/main/java/com/nexaas/challenge/presentation/view/productdetails/ProductDetailsView.kt
@@ -1,0 +1,6 @@
+package com.nexaas.challenge.presentation.view.productdetails
+
+import com.nexaas.challenge.presentation.core.mvp.BaseView
+
+internal interface ProductDetailsView: BaseView {
+}
