--- presentation_layer/src/test/java/com/nexaas/challenge/presentation/view/product/ProductPresenterTest.kt
@@ -1,4 +1,0 @@
-package com.nexaas.challenge.presentation.view.product
-
-class ProductPresenterTest {
-}
