+++ presentation_layer/src/test/java/com/nexaas/challenge/presentation/view/productdetails/ProductPresenterTest.kt
@@ -1,0 +1,4 @@
+package com.nexaas.challenge.presentation.view.productdetails
+
+class ProductPresenterTest {
+}
